+++ NEWS.md
@@ -1,6 +1,6 @@
 # tabularise 0.6.0
 
--   Add {tinytable} table generation for {tabularise} tables as an alternative to {flextable}.
+-   Add {tinytable} table generation for {tabularise} tables as an alternative to {flextable}. Argument `kind =` allows for selecting the table engine to use.
 
 # tabularise 0.5.1